+++ src/capabilities/jobmanager.cpp
@@ -27,6 +27,7 @@
 
 #include "src/global.h"
 
+#include <QtCore/QDateTime>
 #include <QtCore/QObject>
 
 #include "jobmanager.h"
@@ -74,6 +75,9 @@ void JobManager::RunJobs()
         SupraFit::Method method = static_cast<SupraFit::Method>(object["method"].toInt());
         QJsonObject result;
         // qDebug() << object;
+        qint64 t0 = 0, t1 = 0;
+        t0 = QDateTime::currentMSecsSinceEpoch();
+
         switch (method) {
         case SupraFit::Method::WeakenedGridSearch:
             result = RunGridSearch(object);
@@ -81,7 +85,6 @@ void JobManager::RunJobs()
 
         case SupraFit::Method::ModelComparison:
         case SupraFit::Method::FastConfidence:
-
             result = RunModelComparison(object);
             break;
 
@@ -99,6 +102,8 @@ void JobManager::RunJobs()
             result = RunGlobalSearch(object);
             break;
         }
+        t1 = QDateTime::currentMSecsSinceEpoch();
+        emit m_model->Info()->Message(QString("%1 took %2 msecs for %3 in project %4").arg(Method2Name(method)).arg(t1 - t0).arg(m_model->Name()).arg(m_model->ProjectTitle()));
         int index = m_model->UpdateStatistic(result);
         emit ShowResult(method, index);
     }