+++ src/ui/dialogs/thermogram.cpp
@@ -142,12 +142,6 @@ void Thermogram::setUi()
     m_scale->addItem("1");
     m_scale->setEditable(true);
     connect(m_scale, &QComboBox::currentTextChanged, m_scale, [this]() {
-#pragma message("remove me")
-        /*
-        bool ok;
-        qreal scale = m_scale->currentText().toDouble(&ok);
-        if (ok)
-        */
         this->UpdateTable();
     });
     m_injct = new QLineEdit;