+++ Changelog.md
@@ -1,5 +1,7 @@
 # SupraFit 2.0 pre-Alpha
 
+- UI improvments for Results Dialog 
+- Message Dock writes message and warnings
 - option to save specific results from results dialog
 - option for CXO map calculation
 - add option for advanced ui tools